requirements.txt
asgiref==3.6.0
black==23.1.0
btrfsutil==6.1.3
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.3.0
idna==3.4
lxml==4.9.2
Mako==1.2.4
mallard-ducktype==1.0.2
Markdown==3.4.1
MarkupSafe==2.1.2
meson==1.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pycairo==1.23.0
PyGObject==3.42.2
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
TBB==0.2
team==1.0
tomli==2.0.1
urllib3==1.26.14
